Bug fix: relative signal level
The previous code relied on a stable dictionary order, in
the way collections.OrderedDict provides. Instead, modify
our regex to require whitespace after the quality capture,
so only one of "relative" or "absolute" matches.

Look-ahead assertions do not work here:

   \d+(?!/) for 78bla

Matches for "78", as well as "78b", but it's apparently
ambigious what gets captured, even though we expect \d+
to be greedy.